+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * [HexaPDF::Type::AcroForm::Form#each_field] to gracefully handle null values
 * [HexaPDF::Type::AcroForm::AppearanceGenerator] to take an appearance string
   set on a widget instead of a field into account
+* [HexaPDF::Type::AcroForm::ChoiceField] to take PDFs where the /Opt key is set
+  on the widgets into account
 
 
 ## 0.37.1 - 2024-02-05